backend/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
gunicorn==22.0.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
langchain==0.2.11
langchain-google-genai==1.0.7
pytesseract==0.3.10
pillow==10.4.0
pydantic>=2.11.7,<3.0.0
pydantic-settings==2.3.4
python-dotenv==1.0.1
requests==2.32.3
passlib==1.7.4
google-genai==1.44.0
alembic==1.13.2
celery==5.4.0 
redis==5.0.7
gevent==25.9.1
pgvector==0.3.2
sentence-transformers==3.0.1
google-cloud-translate==3.12.1
authlib==1.3.1
itsdangerous==2.2.0
httpx==0.28.1
google-auth
google-auth-oauthlib 
google-auth-httplib2
fastapi-mail==1.4.1
email-validator==2.1.0
pdfplumber==0.11.4
spacy==3.8.2
Jinja2==3.1.4
supabase==2.28.3